backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
asyncpg
alembic
psycopg2-binary
PyJWT
bcrypt
python-multipart
redis~=5.1.0